Title: 세 수의 합과 평균 구하기, Time: 124ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240514/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -4,6 +4,6 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int a = sc.nextInt(), b = sc.nextInt(), c = sc.nextInt();
 
-        System.out.printf("%d \n%.0f",a+b+c, ((double) (a+b+c)/3));
+        System.out.printf("%d \n%d",a+b+c, ((int) (a+b+c)/3));
     }
 }
